move Spf<String> to its own file
